cleanup
https://coveord.atlassian.net/browse/CDX-504
